refactor API requester to be multithreaded
