+++ src/main.rs
@@ -26,7 +26,8 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         log::warn!("Entering Sechat-rs, please be aware this is {pre} SW!");
     }
 
-    let requester = backend::nc_request::NCRequest::new(&config).expect("cannot create NCRequest");
+    let requester = backend::nc_request::nc_requester::NCRequest::new(&config)
+        .expect("cannot create NCRequest");
 
     let backend = backend::nc_talk::NCTalk::new(requester, &config).await?;
     let mut ui: ui::app::App<'_, _> = ui::app::App::new(backend, &config);
